Move traits file
