+++ src/transform/mod.rs
@@ -7,7 +7,6 @@ pub mod collecting;
 mod data_processing;
 pub mod pathogenic_gene_variant_info;
 pub mod strategies;
-pub mod traits;
 pub mod transform_module;
 mod utils;
