+++ src/transform/strategies/age_to_iso8601.rs
@@ -1,7 +1,7 @@
 use crate::extract::contextualized_data_frame::ContextualizedDataFrame;
 use crate::transform::error::StrategyError::MappingError;
 use crate::transform::error::{MappingErrorInfo, PushMappingError, StrategyError};
-use crate::transform::traits::Strategy;
+use crate::transform::strategies::traits::Strategy;
 use log::info;
 
 use crate::extract::contextualized_dataframe_filters::Filter;
@@ -141,7 +141,7 @@ mod tests {
     use crate::extract::contextualized_data_frame::ContextualizedDataFrame;
     use crate::transform::error::{MappingErrorInfo, StrategyError};
     use crate::transform::strategies::age_to_iso8601::AgeToIso8601Strategy;
-    use crate::transform::traits::Strategy;
+    use crate::transform::strategies::traits::Strategy;
     use polars::datatypes::AnyValue;
     use polars::frame::DataFrame;
     use polars::prelude::Column;
